refactor: better define
