+++ packages/api/src/ts/impl.ts
@@ -1,0 +1,34 @@
+import { existsSync } from 'node:fs'
+import path from 'node:path'
+
+export type ResolveTSFileIdImpl = (
+  id: string,
+  importer: string
+) => Promise<string | undefined> | string | undefined
+let resolveTSFileIdImpl: ResolveTSFileIdImpl = resolveTSFileIdNode
+export function resolveTSFileId(id: string, importer: string) {
+  return resolveTSFileIdImpl(id, importer)
+}
+
+/**
+ * @limitation don't node_modules and JavaScript file
+ */
+export function resolveTSFileIdNode(id: string, importer: string) {
+  function tryResolve(id: string, importer: string) {
+    const filePath = path.resolve(importer, '..', id)
+    if (!existsSync(filePath)) return
+    return filePath
+  }
+  return (
+    tryResolve(id, importer) ||
+    tryResolve(`${id}.ts`, importer) ||
+    tryResolve(`${id}.d.ts`, importer) ||
+    tryResolve(`${id}/index`, importer) ||
+    tryResolve(`${id}/index.ts`, importer) ||
+    tryResolve(`${id}/index.d.ts`, importer)
+  )
+}
+
+export function setResolveTSFileIdImpl(impl: ResolveTSFileIdImpl) {
+  resolveTSFileIdImpl = impl
+}
