+++ packages/api/src/vue/props.ts
@@ -21,13 +21,17 @@ import {
   type TSPropertySignature,
   type TSType,
   type TSTypeLiteral,
+  type TSTypeReference,
   type TSUnionType,
   type VariableDeclaration,
 } from '@babel/types'
 import {
   type TSFile,
+  type TSProperties,
   type TSResolvedType,
+  type TSScope,
   isTSExports,
+  resolveMaybeTSUnion,
   resolveTSProperties,
   resolveTSReferencedType,
   resolveTSScope,
@@ -35,6 +39,44 @@ import {
 import { keyToString } from '../utils'
 import { type ASTDefinition, DefinitionKind } from './types'
 import { attachNodeLoc, inferRuntimeType } from './utils'
+
+type BuiltInTypesHandler = Record<
+  string,
+  {
+    handleType(resolved: TSTypeReference): TSType | undefined
+    handleTSProperties?(properties: TSProperties): TSProperties
+  }
+>
+const builtInTypesHandlers: BuiltInTypesHandler = {
+  Partial: {
+    handleType(resolved) {
+      return resolved.typeParameters?.params[0]
+    },
+    handleTSProperties(properties) {
+      for (const prop of Object.values(properties.properties)) {
+        prop.optional = true
+      }
+      return properties
+    },
+  },
+  Required: {
+    handleType(resolved) {
+      return resolved.typeParameters?.params[0]
+    },
+    handleTSProperties(properties) {
+      for (const prop of Object.values(properties.properties)) {
+        prop.optional = false
+      }
+      return properties
+    },
+  },
+  Readonly: {
+    handleType(resolved) {
+      return resolved.typeParameters?.params[0]
+    },
+  },
+  // TODO: pick, omit
+}
 
 export async function handleTSPropsDefinition({
   s,
@@ -193,12 +235,18 @@ export async function handleTSPropsDefinition({
       } else {
         const resolvedType = def.value
         if (resolvedType) {
+          const optional =
+            def.optional ||
+            resolveMaybeTSUnion(resolvedType.ast).some(
+              (t) => t.type === 'TSUndefinedKeyword'
+            )
+
           prop = {
             type: await inferRuntimeType({
               scope: resolvedType.scope || file,
               type: resolvedType.ast,
             }),
-            required: !def.optional,
+            required: !optional,
           }
         } else {
           prop = { type: ['null'], required: false }
@@ -243,145 +291,128 @@ export async function handleTSPropsDefinition({
     withDefaultsAst,
   }
 
-  async function resolveDefinitions(
-    typeDeclRaw: TSResolvedType<TSType>
-  ): Promise<{
-    definitions: TSProps['definitions']
-    definitionsAst: TSProps['definitionsAst']
-  }> {
-    const resolved = await resolveTSReferencedType(typeDeclRaw)
-    if (!resolved || isTSExports(resolved))
-      throw new SyntaxError(`Cannot resolve TS definition.`)
+  async function resolveUnion(definitionsAst: TSUnionType, scope: TSScope) {
+    const unionDefs: TSProps['definitions'][] = []
+    const keys = new Set<string>()
+    for (const type of definitionsAst.types) {
+      const defs = await resolveDefinitions({ type, scope }).then(
+        ({ definitions }) => definitions
+      )
+      Object.keys(defs).map((key) => keys.add(key))
+      unionDefs.push(defs)
+    }
 
-    const { type: definitionsAst, scope } = resolved
-    if (definitionsAst.type === 'TSIntersectionType') {
-      const results: TSProps['definitions'] = {}
-      for (const type of definitionsAst.types) {
-        const defMap = await resolveDefinitions({ type, scope }).then(
-          ({ definitions }) => definitions
-        )
-        for (const [key, def] of Object.entries(defMap)) {
-          const result = results[key]
-          if (!result) {
-            results[key] = def
-            continue
-          }
+    const results: TSProps['definitions'] = {}
+    for (const key of keys) {
+      let optional = false
+      let result: TSPropsMethod | TSPropsProperty | undefined
 
-          if (result.type === 'method' && def.type === 'method') {
-            result.methods.push(...def.methods)
-          } else {
-            results[key] = def
-          }
+      for (const defMap of unionDefs) {
+        const def = defMap[key]
+        if (!def) {
+          optional = true
+          continue
         }
-      }
+        optional ||= def.optional
 
-      return {
-        definitions: results,
-        definitionsAst: buildDefinition({ scope, type: definitionsAst }),
-      }
-    } else if (definitionsAst.type === 'TSUnionType') {
-      const unionDefs: TSProps['definitions'][] = []
-      const keys = new Set<string>()
-      for (const type of definitionsAst.types) {
-        const defs = await resolveDefinitions({ type, scope }).then(
-          ({ definitions }) => definitions
-        )
-        Object.keys(defs).map((key) => keys.add(key))
-        unionDefs.push(defs)
-      }
+        if (!result) {
+          result = def
+          continue
+        }
 
-      const results: TSProps['definitions'] = {}
-      for (const key of keys) {
-        let optional = false
-        let result: TSPropsMethod | TSPropsProperty | undefined
-
-        for (const defMap of unionDefs) {
-          const def = defMap[key]
-          if (!def) {
-            optional = true
+        if (result.type === 'method' && def.type === 'method') {
+          result.methods.push(...def.methods)
+        } else if (result.type === 'property' && def.type === 'property') {
+          if (!def.value) {
             continue
-          }
-          optional ||= def.optional
-
-          if (!result) {
+          } else if (!result.value) {
             result = def
             continue
           }
 
-          if (result.type === 'method' && def.type === 'method') {
-            result.methods.push(...def.methods)
-          } else if (result.type === 'property' && def.type === 'property') {
-            if (!def.value) {
-              continue
-            } else if (!result.value) {
-              result = def
-              continue
-            }
-
-            if (
-              def.value.ast.type === 'TSImportType' ||
-              def.value.ast.type === 'TSDeclareFunction' ||
-              def.value.ast.type === 'TSEnumDeclaration' ||
-              def.value.ast.type === 'TSInterfaceDeclaration' ||
-              def.value.ast.type === 'TSModuleDeclaration' ||
-              result.value.ast.type === 'TSImportType' ||
-              result.value.ast.type === 'TSDeclareFunction' ||
-              result.value.ast.type === 'TSEnumDeclaration' ||
-              result.value.ast.type === 'TSInterfaceDeclaration' ||
-              result.value.ast.type === 'TSModuleDeclaration'
-            ) {
-              // no way!
-              continue
-            }
-
-            if (result.value.ast.type === 'TSUnionType') {
-              result.value.ast.types.push(def.value.ast)
-            } else {
-              // overwrite original to union type
-              result = {
-                type: 'property',
-                value: buildDefinition({
-                  scope,
-                  type: {
-                    type: 'TSUnionType',
-                    types: [result.value.ast, def.value.ast],
-                  } satisfies TSUnionType,
-                }),
-                signature: null as any,
-                optional,
-                addByAPI: false,
-              }
-            }
-          } else {
-            throw new SyntaxError(
-              `Cannot resolve TS definition. Union type contains different types of results.`
-            )
+          if (
+            def.value.ast.type === 'TSImportType' ||
+            def.value.ast.type === 'TSDeclareFunction' ||
+            def.value.ast.type === 'TSEnumDeclaration' ||
+            def.value.ast.type === 'TSInterfaceDeclaration' ||
+            def.value.ast.type === 'TSModuleDeclaration' ||
+            result.value.ast.type === 'TSImportType' ||
+            result.value.ast.type === 'TSDeclareFunction' ||
+            result.value.ast.type === 'TSEnumDeclaration' ||
+            result.value.ast.type === 'TSInterfaceDeclaration' ||
+            result.value.ast.type === 'TSModuleDeclaration'
+          ) {
+            // no way!
+            continue
           }
-        }
 
-        if (result) {
-          results[key] = { ...result, optional }
+          if (result.value.ast.type === 'TSUnionType') {
+            result.value.ast.types.push(def.value.ast)
+          } else {
+            // overwrite original to union type
+            result = {
+              type: 'property',
+              value: buildDefinition({
+                scope,
+                type: {
+                  type: 'TSUnionType',
+                  types: [result.value.ast, def.value.ast],
+                } satisfies TSUnionType,
+              }),
+              signature: null as any,
+              optional,
+              addByAPI: false,
+            }
+          }
+        } else {
+          throw new SyntaxError(
+            `Cannot resolve TS definition. Union type contains different types of results.`
+          )
         }
       }
 
-      return {
-        definitions: results,
-        definitionsAst: buildDefinition({ scope, type: definitionsAst }),
+      if (result) {
+        results[key] = { ...result, optional }
       }
-    } else if (
-      definitionsAst.type !== 'TSInterfaceDeclaration' &&
-      definitionsAst.type !== 'TSTypeLiteral' &&
-      definitionsAst.type !== 'TSMappedType'
-    )
-      throw new SyntaxError(
-        `Cannot resolve TS definition: ${definitionsAst.type}.`
+    }
+
+    return {
+      definitions: results,
+      definitionsAst: buildDefinition({ scope, type: definitionsAst }),
+    }
+  }
+
+  async function resolveIntersection(
+    definitionsAst: TSIntersectionType,
+    scope: TSScope
+  ) {
+    const results: TSProps['definitions'] = {}
+    for (const type of definitionsAst.types) {
+      const defMap = await resolveDefinitions({ type, scope }).then(
+        ({ definitions }) => definitions
       )
+      for (const [key, def] of Object.entries(defMap)) {
+        const result = results[key]
+        if (!result) {
+          results[key] = def
+          continue
+        }
 
-    const properties = await resolveTSProperties({
-      scope,
-      type: definitionsAst,
-    })
+        if (result.type === 'method' && def.type === 'method') {
+          result.methods.push(...def.methods)
+        } else {
+          results[key] = def
+        }
+      }
+    }
 
+    return {
+      definitions: results,
+      definitionsAst: buildDefinition({ scope, type: definitionsAst }),
+    }
+  }
+
+  async function resolveNormal(properties: TSProperties) {
     const definitions: TSProps['definitions'] = {}
     for (const [key, sign] of Object.entries(properties.methods)) {
       const methods = sign.map((sign) => buildDefinition(sign))
@@ -408,8 +439,68 @@ export async function handleTSPropsDefinition({
       }
     }
 
+    return definitions
+  }
+
+  async function resolveDefinitions(
+    typeDeclRaw: TSResolvedType<TSType>
+  ): Promise<{
+    definitions: TSProps['definitions']
+    definitionsAst: TSProps['definitionsAst']
+  }> {
+    let resolved = await resolveTSReferencedType(typeDeclRaw)
+
+    let builtInTypesHandler: BuiltInTypesHandler[string] | undefined
+
+    if (
+      resolved &&
+      !isTSExports(resolved) &&
+      resolved.type.type === 'TSTypeReference' &&
+      resolved.type.typeName.type === 'Identifier'
+    ) {
+      const typeName = resolved.type.typeName.name
+
+      let type: TSType | undefined
+      if (typeName in builtInTypesHandlers) {
+        builtInTypesHandler = builtInTypesHandlers[typeName]
+        type = builtInTypesHandler.handleType(resolved.type)
+      }
+
+      if (type)
+        resolved = await resolveTSReferencedType({
+          type,
+          scope: resolved.scope,
+        })
+    }
+
+    if (!resolved || isTSExports(resolved)) {
+      throw new SyntaxError(`Cannot resolve TS definition.`)
+    }
+
+    const { type: definitionsAst, scope } = resolved
+    if (definitionsAst.type === 'TSIntersectionType') {
+      return resolveIntersection(definitionsAst, scope)
+    } else if (definitionsAst.type === 'TSUnionType') {
+      return resolveUnion(definitionsAst, scope)
+    } else if (
+      definitionsAst.type !== 'TSInterfaceDeclaration' &&
+      definitionsAst.type !== 'TSTypeLiteral' &&
+      definitionsAst.type !== 'TSMappedType'
+    )
+      throw new SyntaxError(
+        `Cannot resolve TS definition: ${definitionsAst.type}.`
+      )
+
+    let properties = await resolveTSProperties({
+      scope,
+      type: definitionsAst,
+    })
+
+    if (builtInTypesHandler?.handleTSProperties)
+      properties = builtInTypesHandler.handleTSProperties(properties)
+
     return {
-      definitions,
+      definitions: await resolveNormal(properties),
       definitionsAst: buildDefinition({ scope, type: definitionsAst }),
     }
   }
@@ -531,6 +622,7 @@ export interface TSProps extends PropsBase {
     | TSIntersectionType
     | TSUnionType
     | TSMappedType
+    | TSTypeReference
   >
 
   /**
