+++ packages/better-define/tests/fixtures/built-in.vue
@@ -1,0 +1,12 @@
+<script setup lang="ts">
+type T = {
+  optional: number
+}
+defineProps<
+  Partial<T> &
+    Required<{
+      required?: boolean
+    }> &
+    Readonly<{ readonly: string }>
+>()
+</script>
