+++ packages/better-define/tests/fixtures/playground.vue
@@ -1,0 +1,6 @@
+<script setup lang="ts">
+declare const a: string
+defineProps<{
+  foo: typeof a
+}>()
+</script>
