+++ packages/better-define/tests/fixtures/undefined.vue
@@ -1,0 +1,5 @@
+<script setup lang="ts">
+defineProps<{
+  foo: undefined | string
+}>()
+</script>
